First pass on preserving IGT between application exits
Adds 2 new variables used to detect and recover IGT after the application exits.

LastCurrentTime is used to see if CurrentTime dropped below it signifying the app has closed.
PreCrashTime is used to store the last known value of CurrentTime and is used in calculating the time to display once recovered or to store when splitting.

Also added a fix to prevent grabbing the current GameTime until StartTime is ready.  This prevents a issue with these new variables, but also seemingly fixes the bug with a quick glitched timer before a stage starts.